+++ torch_xla/experimental/dynamo_mark_sharding.py
@@ -1,32 +1,5 @@
-import torch
-import torch_xla
+import warnings
 
-from typing import Tuple, Union, List, Sequence, Any, Optional, Set
-from torch.library import impl
-from torch_xla.core.xla_model import XLA_LIB
-import torch_xla.distributed.spmd as xs
-
-import ast
-
-XLA_LIB.define(
-    "dynamo_mark_sharding(Tensor input, int[] device_ids, int[] mesh_shape, str axis_names, str partition_spec) -> Tensor"
+warnings.warn(
+    "dynamo_mark_sharding will be auto registered starting 2.5 release, please remove this import"
 )
-
-
-@impl(XLA_LIB, "dynamo_mark_sharding", "XLA")
-def dynamo_mark_sharding_xla(input: torch.Tensor, device_ids: List[int],
-                             mesh_shape: List[int], axis_names: str,
-                             partition_spec: str):
-  mesh_shape_tuple = tuple(mesh_shape)
-  axis_names_eval = ast.literal_eval(axis_names)
-  mesh = xs.Mesh(device_ids, mesh_shape, axis_names_eval)
-  partition_spec_eval = ast.literal_eval(partition_spec)
-  return xs.mark_sharding(input, mesh, partition_spec_eval)
-
-
-@impl(XLA_LIB, "dynamo_mark_sharding", "CompositeExplicitAutograd")
-def dynamo_mark_sharding(input: torch.Tensor, device_ids: List[int],
-                         mesh_shape: List[int], axis_names: str,
-                         partition_spec: str):
-  # Do nothing for non-xla tensor.
-  return input
